+++ components/atoms/Search/search.tsx
@@ -17,7 +17,7 @@ interface SearchProps {
 const suggestionsStyle = {
   "-webkit-line-clamp": "1",
   display: "-webkit-box",
-  "-webkit-box-orient": "vertical"
+  "-webkit-box-orient": "vertical",
 };
 
 const Search = ({
@@ -29,7 +29,7 @@ const Search = ({
   onSearch,
   suggestions,
   onChange,
-  isLoading
+  isLoading,
 }: SearchProps): JSX.Element => {
   const [search, setSearch] = useState(value);
   const [showSuggestions, setShowSuggestions] = useState(false);
@@ -81,19 +81,19 @@ const Search = ({
           }
         }}
         onFocus={() => setShowSuggestions(true)}
-        onBlur={() => setTimeout(() => setShowSuggestions(false), 100)}
+        onBlur={() => setTimeout(() => setShowSuggestions(false), 500)}
       />
 
       {suggestions && suggestions.length > 0 && showSuggestions && (
-        <div className="absolute left-0 z-10 w-full pb-3 bg-white border rounded-lg cursor-pointer shadow-input border-light-slate-6 top-full">
+        <div className="absolute left-0 z-10 w-full pb-1 space-y-1 bg-white border rounded-lg cursor-pointer shadow-input border-light-slate-6 top-full">
           {suggestions.map((suggestion, index) => (
             <div
-              className="px-4 py-2 my-0.5 h-8 break-all  overflow-hidden"
+              className="px-4 py-2 overflow-hidden break-all text-light-slate-9 hover:bg-light-slate-2"
               style={suggestionsStyle}
               key={index}
               onClick={() => handleOnSelect(suggestion)}
             >
-              <span className="pl-5 text-sm text-light-slate-9">{suggestion}</span>
+              <span className="pl-5 text-sm ">{suggestion}</span>
             </div>
           ))}
         </div>