+++ components/molecules/PaginationGotoPage/pagination-goto-page.tsx
@@ -41,7 +41,7 @@ const PaginationGotoPage = ({ page, name, setPage, totalPage }: PaginationGotoPa
         className="min-w-4 w-10 text-center text-sm text-light-slate-10 focus:outline-none border rounded-lg py-1 px-2"
       />
       <button
-        className="text-light-orange-10 text-sm px-4 py-1.5 rounded-lg hover:bg-orange-100 bg-light-orange-3"
+        className="text-light-orange-10 text-sm px-2 py-1.5 rounded-lg hover:bg-orange-100 bg-light-orange-3"
         type="submit"
       >
         Go to page